Increase warning level and remove dangerous ones
- Add explicit nonatomic attribute to properties, since we really don't
need them to be atomic.
- Remove redefinitions of variables
- Remove any implicit typecasts and make them explicit.

+++ Source/DirectoryWatcher.h
@@ -12,7 +12,7 @@
 }
 
 @property (assign) NSObject <DirectoryWatcherDelegate> *delegate;
-@property double latency;
+@property NSUInteger latency;
 
 
 + (id)directoryWatcherWithPath:(NSString *)path;
